requirements.txt
streamlit>=1.31.0
scikit-learn>=1.4.0
python-docx>=1.1.0
PyPDF2>=3.0.0
pandas>=2.2.0
numpy>=1.26.0
nltk>=3.8.1 